Decouple interpreter from core
