+++ src/test/script_P2SH_tests.cpp
@@ -5,6 +5,8 @@
 #include "key.h"
 #include "keystore.h"
 #include "main.h"
+#include "script/interpreter.h"
+#include "script/interpreter_core.h"
 #include "script/script.h"
 #include "script/sign.h"
 #include "wallet_ismine.h"
