+++ src/test/sighash_tests.cpp
@@ -8,6 +8,7 @@
 #include "serialize.h"
 #include "script/script.h"
 #include "script/interpreter.h"
+#include "script/interpreter_core.h"
 #include "util.h"
 #include "version.h"
